Some refactorings that might be nice now that cell_value becomes such an essential part of the library
