+++ src/chrono.rs
@@ -4,8 +4,8 @@
 use chrono::{NaiveDate, NaiveDateTime, NaiveTime};
 use leptos::*;
 
-impl crate::CellValue for NaiveDate {
-    fn render_value(self, options: &crate::RenderOptions) -> impl IntoView {
+impl crate::cell_value::CellValue for NaiveDate {
+    fn render_value(self, options: &crate::cell_value::RenderOptions) -> impl IntoView {
         if let Some(value) = options.format_string.as_ref() {
             view! {
                 <>{self.format(&value).to_string()}</>
@@ -17,8 +17,8 @@ impl crate::CellValue for NaiveDate {
         }
     }
 }
-impl crate::CellValue for NaiveDateTime {
-    fn render_value(self, options: &crate::RenderOptions) -> impl IntoView {
+impl crate::cell_value::CellValue for NaiveDateTime {
+    fn render_value(self, options: &crate::cell_value::RenderOptions) -> impl IntoView {
         if let Some(value) = options.format_string.as_ref() {
             view! {
                 <>{self.format(&value).to_string()}</>
@@ -31,8 +31,8 @@ impl crate::CellValue for NaiveDateTime {
     }
 }
 
-impl crate::CellValue for NaiveTime {
-    fn render_value(self, options: &crate::RenderOptions) -> impl IntoView {
+impl crate::cell_value::CellValue for NaiveTime {
+    fn render_value(self, options: &crate::cell_value::RenderOptions) -> impl IntoView {
         if let Some(value) = options.format_string.as_ref() {
             view! {
                 <>{self.format(&value).to_string()}</>
